requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-WTF>=1.2.0
Flask-Login>=0.6.3
bcrypt>=4.0.0
Werkzeug>=3.0.0
psycopg2-binary>=2.9.0
email-validator>=2.1.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0
